feat: Prepare for proper input implementation
