+++ packages/ui/src/components/Input/Input/Input.tsx
@@ -1,198 +1,13 @@
-/* eslint-disable no-unused-vars */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-
 import React, { ElementRef } from 'react'
+import { useUuid } from '@aviato/hooks'
 import { ComponentProps } from '@stitches/react'
-import { classNames, styled } from '~/theme'
-import { Conditional } from '~/components/Utilities'
+import { StitchedCSS } from '~/theme'
 
-const InputWrapper = styled('div', {
-  position: 'relative',
-  width: '100%',
-  borderRadius: '4px',
+import { BaseInput, BaseInputProps, StyledInput } from './BaseInput'
 
-  '&::after': {
-    content: `''`,
-    display: 'block',
-    position: 'absolute',
-    width: '100%',
-    height: '100%',
-    top: '0',
-    left: '0',
-    pointerEvents: 'none',
-    borderRadius: '4px',
-  },
+const InputStyles: StitchedCSS = {}
 
-  '&.isDisabled': {
-    background: '$OtherDisabledBackground',
-
-    '&::after': {
-      border: '1px solid $OtherDisabledOutline',
-    },
-  },
-
-  '&.isInvalid': {
-    '&::after': {
-      border: '1px solid $ErrorOutline',
-    },
-  },
-
-  variants: {
-    variant: {
-      outlined: {
-        '&::after': {
-          border: '1px solid $OtherInputBorderDefault',
-        },
-
-        '&:hover': {
-          '&::after': {
-            border: '1px solid $OtherInputBorderHover',
-          },
-
-          '&.isActive': {
-            '&::after': {
-              border: '2px solid $OtherInputBorderActive',
-            },
-          },
-        },
-
-        '&.isActive': {
-          '&::after': {
-            border: '2px solid $OtherInputBorderActive',
-          },
-        },
-      },
-
-      filled: {
-        '&::after': {
-          border: '1px solid transparent',
-        },
-
-        '&:not(.isActive):not(.isDisabled)': {
-          background: '$ActionMain',
-
-          '&:hover': {
-            background: '$ActionMainHover',
-          },
-        },
-
-        '&.isActive': {
-          '&::after': {
-            border: '2px solid $OtherInputBorderActive',
-          },
-        },
-
-        '&:hover': {
-          '&.isActive': {
-            '&::after': {
-              border: '2px solid $OtherInputBorderActive',
-            },
-          },
-        },
-      },
-
-      unstyled: {
-        '&.isDisabled': {
-          background: 'none',
-          border: '1px solid transparent',
-        },
-      },
-    },
-  },
-})
-
-export const StyledInput = styled('input', {
-  position: 'relative',
-  display: 'block',
-  height: '36px',
-  minHeight: '36px',
-  lineHeight: '1.55',
-  fontSize: '15px',
-  width: '100%',
-  minWidth: '0px',
-  textAlign: 'left',
-  paddingLeft: '12px',
-  color: '$TextPrimary',
-  background: 'transparent',
-
-  '&:disabled': {
-    background: 'transparent',
-    color: '$OtherDisabledContent',
-
-    '&::placeholder': {
-      color: '$OtherDisabledContent',
-    },
-  },
-
-  '&::placeholder': {
-    color: '$TextSecondary',
-  },
-
-  '&.hasLeftIcon': {
-    paddingLeft: '32px',
-  },
-
-  '&.hasRightIcon': {
-    paddingRight: '32px',
-  },
-})
-
-const IconWrapper = styled('span', {
-  position: 'absolute',
-  top: 0,
-  bottom: 0,
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  pointerEvents: 'none',
-  width: '36px',
-  height: '36px',
-  zIndex: 1,
-  color: '$TextSecondary',
-
-  '&.isActive': {
-    color: '$TextPrimary',
-  },
-
-  '&.isDisabled': {
-    color: '$OtherDisabledContent',
-  },
-
-  variants: {
-    type: {
-      start: {
-        left: 0,
-      },
-      end: {
-        right: 0,
-      },
-    },
-  },
-})
-
-export type InputVariant = 'outlined' | 'filled' | 'unstyled'
-export type InputType =
-  | 'text'
-  | 'password'
-  | 'email'
-  | 'search'
-  | 'tel'
-  | 'url'
-  | 'number'
-
-export interface InputProps {
-  component?: React.ElementType
-  type?: InputType
-  placeholder?: string
-  leftIcon?: React.ReactNode
-  rightIcon?: React.ReactNode
-  variant?: InputVariant
-  disabled?: boolean
-  invalid?: boolean
-  multiline?: boolean
-  maxRows?: number
-  minRows?: number
-}
+export interface InputProps extends BaseInputProps {}
 
 type ForwardProps = ComponentProps<typeof StyledInput> & InputProps
 
@@ -200,52 +15,16 @@ export const Input = React.forwardRef<
   ElementRef<typeof StyledInput>,
   ForwardProps
 >((properties, forwardedRef) => {
-  const {
-    component = 'input',
-    leftIcon = null,
-    rightIcon = null,
-    variant = 'outlined',
-    disabled: isDisabled = false,
-    invalid: isInvalid = false,
-    multiline: isMultiline = false,
-    ...remainingProps
-  } = properties
+  const { ...remainingProps } = properties
 
-  const [isActive, setIsActive] = React.useState(false)
-  const hasLeftIcon = Boolean(leftIcon)
-  const hasRightIcon = Boolean(rightIcon)
-
-  const classes = classNames({
-    hasLeftIcon,
-    hasRightIcon,
-    isActive,
-    isDisabled,
-    isInvalid,
-  })
+  const uuid = useUuid({ prefix: 'input' })
 
   return (
-    <InputWrapper variant={variant} className={classes}>
-      <Conditional test={leftIcon}>
-        <IconWrapper type="start" className={classes}>
-          {leftIcon}
-        </IconWrapper>
-      </Conditional>
-
-      <StyledInput
-        as={component}
-        ref={forwardedRef}
-        className={classes}
-        onFocus={() => setIsActive(true)}
-        onBlur={() => setIsActive(false)}
-        disabled={isDisabled}
-        {...remainingProps}
-      />
-
-      <Conditional test={rightIcon}>
-        <IconWrapper type="end" className={classes}>
-          {rightIcon}
-        </IconWrapper>
-      </Conditional>
-    </InputWrapper>
+    <BaseInput
+      ref={forwardedRef}
+      css={InputStyles}
+      id={uuid}
+      {...remainingProps}
+    />
   )
 })
