+++ packages/ui/src/components/Input/Input/TextField.tsx
@@ -1,47 +1,41 @@
 import React, { ElementRef } from 'react'
 import TextareaAutosize from 'react-textarea-autosize'
 import { useUuid } from '@aviato/hooks'
+import { ComponentProps } from '@stitches/react'
 import { StitchedCSS } from '~/theme'
 
-import { Input, InputProps, StyledInput } from '../Input'
+import { BaseInput, BaseInputProps, StyledInput } from './BaseInput'
 
-export const TextFieldStyles: StitchedCSS = {
+const TextFieldStyles: StitchedCSS = {
   padding: '10px 12px',
   height: 'auto',
   maxHeight: 'auto',
   resize: 'none',
 }
 
-export interface TextFieldProps {
+export interface TextFieldProps extends BaseInputProps {
   autosize?: boolean
   maxRows?: number
   minRows?: number
 }
 
-type ForwardProps = InputProps & TextFieldProps
+type ForwardProps = ComponentProps<typeof StyledInput> & TextFieldProps
 
 export const TextField = React.forwardRef<
   ElementRef<typeof StyledInput>,
   ForwardProps
 >((properties, forwardedRef) => {
-  const {
-    autosize = false,
-    multiline = false,
-    minRows,
-    maxRows,
-    ...remainingProps
-  } = properties
+  const { autosize = false, minRows, maxRows, ...remainingProps } = properties
 
   const uuid = useUuid({ prefix: 'text-field' })
 
   if (autosize) {
     return (
-      <Input
+      <BaseInput
         ref={forwardedRef}
         component={TextareaAutosize}
         css={TextFieldStyles}
         id={uuid}
-        multiline={multiline}
         maxRows={maxRows}
         minRows={minRows}
         {...remainingProps}
@@ -50,12 +44,11 @@ export const TextField = React.forwardRef<
   }
 
   return (
-    <Input
+    <BaseInput
       ref={forwardedRef}
       component="textarea"
       css={TextFieldStyles}
       id={uuid}
-      multiline={multiline}
       {...remainingProps}
     />
   )
