+++ packages/ui/src/components/Input/Input/index.ts
@@ -1,1 +1,3 @@
 export * from './Input'
+export * from './TextField'
+export * from './types'
