+++ packages/ui/src/components/Input/Input/types.ts
@@ -1,0 +1,9 @@
+export type InputVariant = 'outlined' | 'filled' | 'unstyled'
+export type InputType =
+  | 'text'
+  | 'password'
+  | 'email'
+  | 'search'
+  | 'tel'
+  | 'url'
+  | 'number'
