--- packages/ui/src/components/Input/TextField/index.ts
@@ -1,1 +1,0 @@
-export * from './TextField'
