+++ packages/ui/src/components/Input/index.ts
@@ -4,4 +4,3 @@ export * from './Input'
 export * from './RadioGroup'
 export * from './Slider'
 export * from './Switch'
-export * from './TextField'
